feat: small refactor, secure delete space

+++ server/api/space/space_service.go
@@ -1,6 +1,9 @@
 package space
 
 import (
+	"errors"
+	"log"
+
 	"github.com/dev-oleksandrv/db"
 	"github.com/dev-oleksandrv/repository"
 )
@@ -21,6 +24,24 @@ func (s *SpaceService) GetAllSpacesByUserID(userID int) ([]db.Space, error) {
 	return s.spaceRepo.GetAllByUserID(userID) 
 }
 
-func (s *SpaceService) DeleteSpace(id uint) error {
-	return s.spaceRepo.Delete(id)
+func (s *SpaceService) DeleteSpace(userID, id int) error {
+	isAssociated, err := s.spaceRepo.IsUserAssociatedWithSpace(userID, id)
+	
+	if err != nil {
+		log.Printf("Error checking association: %v", err)
+		return err
+	}
+
+	if !isAssociated {
+		log.Printf("User is not authorized to delete this space")
+		return errors.New("user is not authorized to delete this space")
+	}
+
+	err = s.spaceRepo.Delete(id)
+	if err != nil {
+		log.Printf("Error deleting space: %v", err)
+		return err
+	}
+	
+	return nil
 }
